Create hash2username.py
I have dumped the NTDS.dit file and used hashcat to crack multiple hashes.  Then you have to tie that hash back to a username.  Hashcat does have a "--show --username" function (thanks hashcat developers).  Ive had it not work properly all the time.  This is a quick and dirty script to tie the username back to a cracked hash.